requirements.txt
streamlit==1.30.0
deepface==0.0.79
opencv-python-headless==4.8.0.74
tensorflow==2.12.0
numpy==1.23.5
pandas==2.2.2
Pillow==10.0.0